+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.6.0]
+
 ### Added
 
 - vscode command resolution `${command:<...>}`:
@@ -15,7 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - source file resolution
 
-## [3.5.3]
+## [3.5.3] - 2020-10-30
 
 ### Fixed